refactor(bin): extract text edit from help viewer
I will use this as the starting point for my assembler pane.
