+++ bin/gui/textedit/linenumbers.cpp
