+++ bin/gui/textedit/linenumbers.h
