+++ bin/gui/textedit/registration.hpp
@@ -13,11 +13,24 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
+
 #pragma once
 #include <QQmlApplicationEngine>
 #include <QtCore>
 #include "commands/gui.hpp"
 
-namespace helpview {
+class QTextDocument;
+class QQuickTextDocument;
+class BlockFinder : public QObject {
+  Q_OBJECT
+  QTextDocument *_doc = nullptr;
+
+public:
+  BlockFinder(QObject *parent = nullptr);
+  Q_INVOKABLE int find_pos(int pos);
+  Q_INVOKABLE void set_document(QQuickTextDocument *doc);
+};
+
+namespace textedit {
 void registerTypes(QQmlApplicationEngine &engine);
 }
